+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -10,8 +10,8 @@
   GetDefaultResource().
 * Modified the data type of Headers option to string; Added a private method
   GetMetadataFromHeaders to create a metadata object from the headers passed as
-  a string in the options; Added a new option called TimeOut for computing the
-  `deadline` to be used by gRPC client for `Export`
+  a string in the options; Added a new option called TimeoutMilliseconds for
+  computing the `deadline` to be used by gRPC client for `Export`
   ([#1781](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1781))
 
 ## 1.0.0-rc2