+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -8,10 +8,12 @@
 * Changed default port for OTLP Exporter from 55680 to 4317
 * Default ServiceName, if not found in Resource, is obtained from SDK using
   GetDefaultResource().
-* Modified the data type of Headers option to string; Added a new option called
-  TimeoutMilliseconds for computing the `deadline` to be used by gRPC client for
-  `Export`
+* Modified the data type of `Headers` option to string; Added a new option
+  called `TimeoutMilliseconds` for computing the `deadline` to be used by gRPC
+  client for `Export`
   ([#1781](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1781))
+* Modified the data type of `Endpoint` option to string for netstandard2.1
+  ([#1786](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1786))
 
 ## 1.0.0-rc2